ns-o-ran_ns3/requirements.txt
numpy
torch
pandas
statsd
joblib
PyYaml
influxdb
watchdog
statsd-tags
scikit-learn
